remove duplication in test helper classes
